Add `Body::from_stream`
This replaces `StreamBody` with `Body::from_stream`.

Requires `axum-core` to depend on `pin-project-lite` but that is already
pulled in by `pin-project-lite`.
