+++ axum-extra/src/json_lines.rs
@@ -2,7 +2,7 @@
 
 use axum::{
     async_trait,
-    body::{Body, StreamBody},
+    body::Body,
     extract::FromRequest,
     response::{IntoResponse, Response},
     BoxError,
@@ -166,7 +166,7 @@ where
             buf.write_all(b"\n")?;
             Ok::<_, BoxError>(buf.into_inner().freeze())
         });
-        let stream = StreamBody::new(stream);
+        let stream = Body::from_stream(stream);
 
         // there is no consensus around mime type yet
         // https://github.com/wardi/jsonlines/issues/36
