+++ axum/src/body/mod.rs
@@ -1,9 +1,5 @@
 //! HTTP body utilities.
 
-mod stream_body;
-
-pub use self::stream_body::StreamBody;
-
 #[doc(no_inline)]
 pub use http_body::{Body as HttpBody, Empty, Full};
