+++ axum/src/test_helpers/mod.rs
@@ -7,6 +7,5 @@ pub(crate) use self::test_client::*;
 
 pub(crate) fn assert_send<T: Send>() {}
 pub(crate) fn assert_sync<T: Sync>() {}
-pub(crate) fn assert_unpin<T: Unpin>() {}
 
 pub(crate) struct NotSendSync(*const ());
